Create a `FlightController` struct, parameterized by a generic `I2CDevice`
